pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tdf2lance"
dynamic = ["version"]
description = 'Experimental conversion of bruker tdf files to lance files.'
readme = "README.md"
requires-python = ">=3.8"
license = "Apache 2.0"
keywords = []
authors = [
  { name = "J. Sebastian Paez", email = "jspaezp@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
]
dependencies = [
  "click",
  "pylance",
  "duckdb",
  "alphatims",
  "tqdm",
  "loguru",
  "pyarrow",
  "scikit-learn",
  "scipy",
  "psims",
]

[project.urls]
Documentation = "https://github.com/jspaezp/tdf2lance#readme"
Issues = "https://github.com/jspaezp/tdf2lance/issues"
Source = "https://github.com/jspaezp/tdf2lance"

[project.scripts]
tdf2lance = "tdf2lance.cli:tdf2lance"

[tool.hatch.version]
path = "src/tdf2lance/__about__.py"

[tool.hatch.envs.default]
dependencies = [
  "coverage[toml]>=6.5",
  "pytest",
  "scalene", # pretty decent profiler
  "ipykernel",
  "matplotlib",
  "streamlit",
  "optuna",
  "optuna-dashboard",
]
[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = [
  "- coverage combine",
  "coverage report",
]
cov = [
  "test-cov",
  "cov-report",
]

[[tool.hatch.envs.all.matrix]]
python = ["3.8", "3.9", "3.10", "3.11"]

[tool.hatch.envs.lint]
detached = true
dependencies = [
  "black>=23.1.0",
  "ruff>=0.0.243",
]
[tool.hatch.envs.lint.scripts]
style = [
  "ruff {args:.}",
  "black --check --diff {args:.}",
]
fmt = [
  "black {args:.}",
  "ruff --fix {args:.}",
  "style",
]
all = [
  "style",
]

[tool.black]
target-version = ["py38"]
line-length = 88
skip-string-normalization = true

[tool.ruff]
target-version = "py38"
line-length = 88
select = [
  "A",
  "ARG",
  "B",
  "C",
  "DTZ",
  "E",
  "EM",
  "F",
  "FBT",
  "I",
  "ICN",
  "ISC",
  "N",
  "PLC",
  "PLE",
  "PLR",
  "PLW",
  "Q",
  "RUF",
  "S",
  "T",
  "TID",
  "UP",
  "W",
  "YTT",
]
ignore = [
  # Allow non-abstract empty methods in abstract base classes
  "B027",
  # Allow boolean positional values in function calls, like `dict.get(... True)`
  "FBT003",
  # Ignore checks for possible passwords
  "S105", "S106", "S107",
  # Ignore complexity
  "C901", "PLR0911", "PLR0912", "PLR0913", "PLR0915",
]
unfixable = [
  # Don't touch unused imports
  "F401",
]

[tool.ruff.isort]
known-first-party = ["tdf2lance"]

[tool.ruff.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.per-file-ignores]
# Tests can use magic values, assertions, and relative imports
"tests/**/*" = ["PLR2004", "S101", "TID252"]

[tool.coverage.run]
source_pkgs = ["tdf2lance", "tests"]
branch = true
parallel = true
omit = [
  "src/tdf2lance/__about__.py",
]

[tool.coverage.paths]
tdf2lance = ["src/tdf2lance", "*/tdf2lance/src/tdf2lance"]
tests = ["tests", "*/tdf2lance/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
